Added condition to prevent multiple API calls
Depending on the speed of the site's database, multiple API calls may be made for the same short URL request while a post is being published. Causes no error on the site, but does create duplicate links on Rebrandly, all but one go unused. Added conditional to check for a saved post meta short URL key before the API call to prevent duplication.